+++ app/handlers_flightrecord.go
@@ -182,3 +182,14 @@ func (app *application) HandlerApiTrackLogReset(w http.ResponseWriter, r *http.R
 
 	app.writeOkResponse(w, "Track Log reset")
 }
+
+// HandlerApiFlightRecordTags returns all unique flight records tags
+func (app *application) HandlerApiFlightRecordTags(w http.ResponseWriter, r *http.Request) {
+	tags, err := app.db.GetFlightRecordsTags()
+	if err != nil {
+		app.handleError(w, err)
+		return
+	}
+
+	app.writeJSON(w, http.StatusOK, tags)
+}