+++ internal/driver/db_structure.go
@@ -1,7 +1,7 @@
 package driver
 
 var (
-	schemaVersion = "17"
+	schemaVersion = "18"
 
 	UUID      = ColumnType{SQLite: "TEXT", MySQL: "VARCHAR(36)"}
 	DateTime  = ColumnType{SQLite: "TEXT", MySQL: "VARCHAR(32)"}
@@ -48,6 +48,7 @@ var logbookTable = NewTable("logbook", "uuid", UUID,
 		{Name: "distance", Type: Real},
 		{Name: "track", Type: Blob},
 		{Name: "custom_fields", Type: FullText},
+		{Name: "tags", Type: BigText, Properties: "NOT NULL DEFAULT ''"},
 	})
 
 var airportsTable = NewTable("airports", "icao", SmallText,
@@ -170,7 +171,8 @@ var logbookView = NewView("logbook_view",
 				instructor_time, sim_type, sim_time, pic_name, remarks,
 				IFNULL(distance, 0) distance, track, IFNULL(custom_fields, '{}') as custom_fields,
 				CASE WHEN track IS NULL THEN 0 ELSE 1 END AS has_track,
-				IFNULL(ac.cnt, 0) AS attachments_count
+				IFNULL(ac.cnt, 0) AS attachments_count,
+				IFNULL(tags, '') AS tags
 			FROM logbook
 			LEFT JOIN (
   				SELECT record_id, COUNT(*) AS cnt FROM attachments GROUP BY record_id
@@ -187,7 +189,8 @@ var logbookView = NewView("logbook_view",
 				instructor_time, sim_type, sim_time, pic_name, remarks,
 				IFNULL(distance, 0) distance, track, IFNULL(custom_fields, '{}') as custom_fields,
 				CASE WHEN track IS NULL THEN 0 ELSE 1 END AS has_track,
-				IFNULL(ac.cnt, 0) AS attachments_count
+				IFNULL(ac.cnt, 0) AS attachments_count,
+				IFNULL(tags, '') AS tags
 			FROM logbook
 			LEFT JOIN (
   				SELECT record_id, COUNT(*) AS cnt FROM attachments GROUP BY record_id